Adding the ability to copy the data when creating a MemorySource.  This fixed a bug in the Java bindings where a MemorySource was created but the memory wasn't guaranteed to still be around by the time it was used by the writer to write it out.
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@1038 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ java/nitf/src/jni/source/nitf_ImageSource.c
@@ -104,7 +104,7 @@ JNIEXPORT jobjectArray JNICALL Java_nitf_ImageSource_getBandSources
     end = nitf_List_end(imageSource->bandSources);
     while (nitf_ListIterator_notEqualTo(&iter, &end))
     {
-        address = (nitf_BandSource *)nitf_ListIterator_get(&iter);
+        address = (jlong)((nitf_BandSource *)nitf_ListIterator_get(&iter));
         element = (*env)->CallStaticObjectMethod(env, bandSourceClass,
                 methodID, address);
         (*env)->SetObjectArrayElement(env, sources, index++, element);
